modify: constant 파일 수정
game-area의 상수들을 같은 room 페이지의 상수들과 함께 관리하기 위해 room 페이지 상수로 통합
